feat: GitHub Actions release action for monorepo (#3)
* feat: GitHub Actions release action for monorepo

A new GitHub TypeScript Action to help us make releases for the
Monorepo. Work in progress.

Uses a combination of turborepo, git, and github rest api to figure out
what the release should be. Works as follows:

1. Loads inputs and tokens for the action and sets up octokit (to use the
   github rest api).

2. Finds the previous release's sha if any. If no release found
   defaults to the previous sha. Should usually not happen.

3. Logs all git commits between release and current sha (the one being
   deployed). List contains sha and subject (heading of git message)

4. Starts processing each commit using turbo to figure out if the
   package/workspace being released was affected by the commit. If
   affected it includes the commit in the list of messages.

5. It processes each commit and groups them by the conventional commit
   type. This allows us to present them more nicely.

6. Finally it builds the release body and creates the release using the
   github api. It also outputs the release title and release body as
   github action outputs. This allows workflows to for exampel post a
   slack message.

This work has not yet been tested in practice. But need some feedback
before we begin testing it.

* Refactor logic into helper functions, add tests

Refactors most of the logic out from main module's run function into
separate helper functions in a helpers module. The helper functions
accept the github context, octokit (rest client) and similar as
parameters. Avoiding access of global github values makes these helper
functions easier to test and mock.

Add tests for all helper functions and the main module for 100% test
coverage. This should make it easier to make changes without breaking
the code, and also informs developers about the expected behaviour of
the various helper functions.

* Allow non-monorepos

If we're not in a monorepo we allow commits to be included in release as
long as they are conventional commits. This is particularly useful when
testing.

* Make from and to shas inputs to action

Instead of this action trying to figure out the commit range from the
previous deployment it can get the from and to commit shas (or other ref
names) for the range to create the release for.

We can create a separate action to find the from commit.

* git log the range with the start commit as exclusive.

Trying to include the from sha in the commit history proved difficult, especially when the commit is the initial repository commit as is the case in the tests. Better for the main use case if the from commit is treated as exclusive in any case.

* Do git checkout to optional branch

This allows creating releases for different branches than the one the
job is running for.

Co-authored-by: Sindre Bøyum <boyum@users.noreply.github.com>