feat: Initial implementation of active deployment action (#1)
* feat: Initial implementation of active deployment action

Implement action that finds the nth most recent active deploy in a
GitHub repository. Defaults to finding the first deployment. If nth is 2
it will find the second most recent deployment.

Allows users to override which repository (owner and repo) to fetch
deployments from. Defaults to the one in the github.context.

Will page through deployments with a slight politeness delay to avoid
spamming GitHub API.

* Improvements, lint fixes, docs

Add default value for github-token so users don't need to specify it.
Only if they require a custom token other than the one included in
GitHub actions do they need to provide it explicitly. Make all inputs
required, but provide default values for all but the evironment.